+++ test/test_retryable_writes.py
@@ -80,6 +80,7 @@ class TestAllScenarios(SpecRunner):
 
 def create_test(scenario_def, test, name):
     @client_context.require_test_commands
+    @client_context.require_no_mmap
     def run_scenario(self):
         self.run_scenario(scenario_def, test)
 
@@ -176,9 +177,41 @@ class IgnoreDeprecationsTest(IntegrationTest):
         super(IgnoreDeprecationsTest, cls).tearDownClass()
 
 
+class TestRetryableWritesMMAPv1(IgnoreDeprecationsTest):
+
+    @classmethod
+    def setUpClass(cls):
+        super(TestRetryableWritesMMAPv1, cls).setUpClass()
+        # Speed up the tests by decreasing the heartbeat frequency.
+        cls.knobs = client_knobs(heartbeat_frequency=0.1,
+                                 min_heartbeat_interval=0.1)
+        cls.knobs.enable()
+        cls.client = rs_or_single_client(retryWrites=True)
+        cls.db = cls.client.pymongo_test
+
+    @classmethod
+    def tearDownClass(cls):
+        cls.knobs.disable()
+
+    @client_context.require_version_min(3, 5)
+    @client_context.require_no_standalone
+    def test_actionable_error_message(self):
+        if client_context.storage_engine != 'mmapv1':
+            raise SkipTest('This cluster is not running MMAPv1')
+
+        expected_msg = ("This MongoDB deployment does not support retryable "
+                        "writes. Please add retryWrites=false to your "
+                        "connection string.")
+        for method, args, kwargs in retryable_single_statement_ops(
+                self.db.retryable_write_test):
+            with self.assertRaisesRegex(OperationFailure, expected_msg):
+                method(*args, **kwargs)
+
+
 class TestRetryableWrites(IgnoreDeprecationsTest):
 
     @classmethod
+    @client_context.require_no_mmap
     def setUpClass(cls):
         super(TestRetryableWrites, cls).setUpClass()
         # Speed up the tests by decreasing the heartbeat frequency.
@@ -429,6 +462,7 @@ class TestRetryableWrites(IgnoreDeprecationsTest):
 class TestRetryableWritesTxnNumber(IgnoreDeprecationsTest):
     @client_context.require_version_min(3, 6)
     @client_context.require_replica_set
+    @client_context.require_no_mmap
     def test_increment_transaction_id_without_sending_command(self):
         """Test that the txnNumber field is properly incremented, even when
         the first attempt fails before sending the command.