+++ test/test_session.py
@@ -1057,6 +1057,7 @@ class TestCausalConsistency(unittest.TestCase):
             self.assertIsNone(act)
 
     @client_context.require_no_standalone
+    @client_context.require_no_mmap
     def test_read_concern(self):
         with self.client.start_session(causal_consistency=True) as s:
             coll = self.client.pymongo_test.test