+++ doc/changelog.rst
@@ -68,6 +68,17 @@ Version 3.9 adds support for MongoDB 4.2. Highlights include:
 - Connections now survive primary step-down. Applications should expect less
   socket connection turnover during replica set elections.
 
+Unavoidable breaking changes:
+
+- Applications that use MongoDB with the MMAPv1 storage engine must now
+  explicitly disable retryable writes via the connection string
+  (e.g. ``MongoClient("mongodb://my.mongodb.cluster/db?retryWrites=false")``) or
+  the :class:`~pymongo.mongo_client.MongoClient` constructor's keyword argument
+  (e.g. ``MongoClient("mongodb://my.mongodb.cluster/db", retryWrites=False)``)
+  to avoid running into :class:`~pymongo.errors.OperationFailure` exceptions
+  during write operations. The MMAPv1 storage engine is deprecated and does
+  not support retryable writes which are now turned on by default.
+
 .. _URI options specification: https://github.com/mongodb/specifications/blob/master/source/uri-options/uri-options.rst